fix(F1): Add some changes related to state bag and mapper
