+++ Src/Core/F1/Presentation/Filters/SetStateBag/F1StateBag.cs
@@ -1,3 +1,8 @@
 namespace F1.Presentation.Filters.SetStateBag;
 
-public sealed class F1StateBag { }
+public sealed class F1StateBag
+{
+    public F1Request HttpRequest { get; set; }
+
+    public F1Response HttpResponse { get; set; }
+}
