+++ Src/Core/F1/Presentation/Filters/Validation/F1ValidationFilter.cs
@@ -1,8 +1,8 @@
-using System.Linq;
 using System.Net.Mime;
 using System.Text.Json;
 using System.Threading.Tasks;
 using F1.Common;
+using F1.Presentation.Filters.SetStateBag;
 using FluentValidation;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Filters;
@@ -23,25 +23,8 @@ public sealed class F1ValidationFilter : IAsyncActionFilter
         ActionExecutionDelegate next
     )
     {
-        var doesRequestExist = context.ActionArguments.Any(argument =>
-            argument.Key.Equals(F1Constant.REQUEST_ARGUMENT_NAME)
-        );
-
-        if (!doesRequestExist)
-        {
-            context.Result = new ContentResult
-            {
-                StatusCode = F1Constant.DefaultResponse.Http.VALIDATION_FAILED.HttpCode,
-                Content = JsonSerializer.Serialize(
-                    F1Constant.DefaultResponse.Http.VALIDATION_FAILED
-                ),
-                ContentType = MediaTypeNames.Application.Json,
-            };
-
-            return;
-        }
-
-        var request = context.ActionArguments[F1Constant.REQUEST_ARGUMENT_NAME] as F1Request;
+        var stateBag = context.HttpContext.Items[nameof(F1StateBag)] as F1StateBag;
+        var request = stateBag.HttpRequest;
 
         var result = await _validator.ValidateAsync(request);
         if (!result.IsValid)
